Extract post-turn PR transition test support
